Merge with Duck Programming Language 1.00.

+++ duck/types.c
@@ -3,10 +3,10 @@
 #include "types.h"
 
 /* cast to float */
-double TypeFloat(VALUE value)
+long double TypeFloat(VALUE value)
 {
     if (value.type == VAL_PRIMITIVE) {
-        return (double)value.data.primitive;
+        return (long double)value.data.primitive;
     } else if (value.type == VAL_FLOATING_POINT) {
         return value.data.floatp;
     } else if (value.type == VAL_STRING) {
@@ -17,12 +17,12 @@ double TypeFloat(VALUE value)
 }
 
 /* cast to int */
-int TypeInt(VALUE value)
+int64 TypeInt(VALUE value)
 {
     if (value.type == VAL_PRIMITIVE) {
         return value.data.primitive;
     } else if (value.type == VAL_FLOATING_POINT) {
-        return (int)value.data.floatp;
+        return (int64)value.data.floatp;
     } else if (value.type == VAL_STRING) {
         return atoi(value.data.string);
     } else {
